+++ HISTORY.md
@@ -1,4 +1,9 @@
 # 0.1.0
+
+`2021-07-31`
+### Changed
+    - Change views to fix PreviewProvider.
+    
 `2021-07-30`
 ### Added
     - Add AlbumGridView.